More cleanup, more docstrings, updated test baseline and a few new tests. Still working on the cleanup.
Also one syntax change: integers are now written as "int<n>" with n
being the number of bits they are wide; this used to be "int:n". The
new syntax will be used by other parameterizable data types as well,
such as for example "list<int>". There are still the short-cuts
int8, int16, int32, and int64.