Updating documentation that describes how to add a new data type.
Robin: I had to change :download: to :file: in the file include.rst in order to
be able to compile the documentation. The :download: primitive was not
recognized by the current sphinx version. You might want to change it back to
something else if you wanted to convey a different meaning.