A number of updates:
- The fmt() function now support %d/%f/%p specifiers with the
obvious meaning. It performans the conversion for the two former via
the type-info's to_{int64,double} so that all types implementing any
of those are automatically supported. 

- All variables are now automatically intialized with default
values---no more unintialized variable values. The default is
defined on a per-type basis with codegens' new @defaultInitValue()
decorator, which all ValueTypes must provides. In the future,
explicit intialization of variables with arbitrary constants will
also be supported. (As an implementation detail, the string data
type now treats null pointers as the empty string, that simplifies
initialization of string values.)

- Bugfix for passing tuples to C: now the individual componets  are
passing through the LLVM-to-C conversion as well, which is crucial
for recursive tuples to be passed correctly. 

- Other smaller bugfixex

- New and extended tests