- A "bytes" data type, suitable for storing and operating on large amounts of raw data.
- Introduction of "operators", which are a new kind of instruction
not bound to a particular type but overloaded for different types.
The first operator is "new", a generic allocate instruction. Rather
than "struct.new A" (with A being the struct type), one now writes 
"new A", and likewise for other types. Structs and bytes have
already been adjusted to use the new operator; channels not yet. 

See "instructions/operators.py" for an example how operators are
defined; and search for "operator" in codegen/bytes.py for an
example of a type-specific implemention.

- Extensions/fixes to strings.

- Note that I'm switching to a new layout of "hilti_intern.h":
rather than putting all the prototypes in there, I'm splitting the
file into individual subfiles which are recursively included. So
far, there's only bytes.h but I'm planing to move the other
functions into their own files in a similar way. 

- Beginnings of a generic iterator interface which will allow to
iterate over different kinds of containers in a transparent fashion.
Bytes will support that interface, and the libhilti functions are
already there, but the compiler can't generate code for it yet. 

- New tests.