Corrected two race conditions in libhilti.
The first was a potential race condition in the sleeping loop of worker threads. I had theorized that this race condition was the cause for the incorrect behavior of the multithreaded tests. In fact, although it was good to locate and correct this race condition, it was not causing the tests to fail.

The second, and the source of the wrong behavior on the multithreaded tests, was the way that the print function is implemented. Because strings were output using several write operations, printing a string in HILTI was not atomic and race conditions could occur that could corrupt the output, mixing together strings or preventing some characters from being printed at all. This has been corrected with the use of flockfile and funlockfile, which are used to acquire and release the internal locks for file streams such as stdout. Now the HILTI print statement is atomic and the tests show correct behavior.