Retried enabled tail-call optimization in codegen, and it now seems to working, with results identical to what I got by doing it by hand. I'm not sure what the problem was the last time, but it has been resolved; all tests that normally pass are passing. To actually be effective, this requires your copy of llvm-py to be patched with the patch you can get here:
http://code.google.com/p/llvm-py/issues/detail?id=19

Note that all the sources of excess memory usage are not resolved by enabling this. For my sieve of erastothenes demo, pre-TCO I was able to calculate roughly the first 10k primes; much higher than that and the program ran out of memory. With TCO I can calculate roughly the first 50k primes before the program runs out of memory. At least one remaining source of memory consumption is that a new exception frame is allocated every time thread.schedule is used; this memory is never freed. It would be good to come up with some sort of solution to that problem, although maybe it's enough to leave it for the garbage collector to solve.