New "overlay" data type. This one is a bit more complex: it allows to specify the binary layput of a chunk of data to then transparently unpack its individual fields.
Example: (note that I'm not sure yet what the right syntax for
defining an overlay is; but this will do for now):

    overlay MyOverlay {
        f1: int<32> at 0 unpack with Hilti::Packed::Int32Big,
        f2: int<32> at 4 unpack with Hilti::Packed::Int32Big,
        f3: int<32> after f2 unpack with Hilti::Packed::Int32Big,
        f4: int<32> after f3 unpack with Hilti::Packed::Int32Big
    }

This defines 4 subsequent int32 integers in big-endian byte order,
which will be accessible by the names f1-f4. f1 starts at offset 0,
and f2 starts at offset 4. f3 starts right after f2, and f4 starts
right after f3. The definition of f3 and f4 is a bit pointless here
because f1 and f2 are fixed size so we could have had specified
their offsets directly as well. Generally, however, fields can have
non-fixed sizes and then the "after" keyword will ensure that
subsequent fields are still found correctly. 

Once defined, such an overlay can be "attached" to a bytes object
(more precisely, to some position in a bytes object):

    local ref<bytes> b
    local iterator<bytes> i

    b = b"\x00\x01\x02\x03\x04\x05\x06\x07\x08\x09\x0a\x0b\x0c\x0d\x0e\x0f"
    i = bytes.begin b
    
    local MyOverlay o
    overlay.attach o i

Finally, we can then can access the fields:

    local int<32> i32
    i32 = overlay.get o "f3"
       
Overlays live on the stack (not the heap) and so overhead is pretty
low.

(This is checkpoint commit, the overlay implementation still needs a
bit of tuning, better error checking, and more testing.)