- Overlays now check whether they are attached
- Overlays now make used of already computed positions.

- Slighlt changed syntac for the unpack operator: the two iterators
are combined into a tuple (begin,end). That makes space for an
additional, optional argument for passing more information to the
unpack implemetnation.

- alloca()s are factored out into a CodeGen.llvmAlloc() function.
The function always allocates the stack variables inside the entry
block of the current LLVM fucntion, which is the only place where
LLVM's mem2reg optimization looks for them. 

- Overlay tests. More to come.