Infrastructure work:
- Moving to LLVM 2.6 (pre-release 2 at the moment), plus updated
Python bindings (which require a patch to work with 2.6)

- Updates for Snow Leopard. 

- Replacing llvm-gcc with clang. All C code is now compiled with
clang. 

- Because it's getting tedious to install, and maintain, all the
dependencies separately, there is a now a tar-ball available for
download which comes with everything required (LLVM, clang, PLY, etc.)
See hilti/README for details. 

- All this is only tested on Snow Leopard at the moment, and I've
disabled the auto-testers until I get a chance to make sure it runs
fine on the other system as well.