Major rework of the hilti.printer submodule: it now prints an HILTI AST in a form parseable by the HILTI parser. The result is that "hiltic -l" now outputs a representation of its input that is equivalent to the original program.
The test-suite has been extended to verify that this actually works
for all the *.hlt files. To trigger these tests, at the moment one
has to manually run tests/run-tester with an additional "-F printer"
argument. In addition to all the standard tests, this will then run
hiltic two more times with "-p": first on the original input file
and then with the output of the first run. Both should produce the
same result, and if that's indeed the case the normal test is then
performed with the output.

TODO: Not all tests pass this yet which is why this not yet part of
the default "make" in tests/. 

Additional changes:

- The "tester" gets support for running a test a second time with
first passing the input through a user-defined filter program. We
use that for the above to run with the test with the "hiltic -p"
output.

- The "tester" also get its own configuration file, tester.cfg per
default. 

- HILTI AST notes now can have comments associated with them, which
can be printed out via hilti.printer.

- The HILTI parser associates comments in the input with subsequent
entities, keeping them internally in the AST nodes. 

   TODO: That doesn't work quite right yet everywhere. 
   
- Various bug fixes and general cleanup.