Parsers can now be connected to sinks based on their MIME type.
Example:

    export type Main = unit {

        [...]

        var data: sink;

        on %init {
            self.data.connect_mime_type("application/bar");
            }
    };

    export type Sub = unit {
        %mimetype = "application/foo";
        %mimetype = "application/bar";

        s1: bytes &length=2;
        s2: bytes &length=8;
    };

    export type Sub2 = unit {
        %mimetype = "application/bar";

        s: bytes &eod;
    };

In this example, an instance of both Sub and Sub2 will be connected
to Main's sink. As the example shows, a parser can define more than
one MIME type. It can also use a wildcard for the subtype, e.g.,
"application/*".

Types instantiated this way can use "self.mime_type()" to get the
MIME type that was used.

Note that all this works only for *exported* types.

There's also a new debug stream "binpac-sinks" that logs
connection/disconnection of parsers to sinks, and also the data that
is written into sinks.