A &chunked attribute for bytes data, part 1.
When having &chunked, a bytes object will not be fully buffered when
being parsed, as it normally would be, but data is passed on
chunk-wise as it comes in. Example:

    export type test = unit {
       a: bytes &eod &chunked {
           print self.a;
           }
       };

    > echo 1234567890 | ./a.out -i 3 -p Mini::test
    123
    456
    789
    0\x0a

    (Recall that "-i 3" passes data in in pieces of 3 bytes each.
    Without that, the output would be simply "1234567890\x0".)

&chunked also applies to attached sinks.

Todo: What's still missing is a capability to identify the first and
last chunks in hooks, so that potentual initialization/cleanup can be
triggered.