Removed redundant match_state structure
State functions carry state as parameters, and
delegate saving them to the match_state structure
to an external function.

This is done to avoid having to create a mirror
LLVM match_state structure, and to avoid fragile direct access via member offsets.

Instead, the match_state structure is
passed through the LLVM code as an unknown
"opaque" pointer, and is handled correctly in
an external C function.