Bringing the compilation cache back.
In addition to caching *.ll files, we're now also using a new LLVM 3.3
capability to cache the native code for avoiding the JIT process:

    > rm -rf .cache
    > time bro -r ssh-single-conn.trace ssh-banner.bro
    SSH banner, [orig_h=192.150.186.169, orig_p=49244/tcp, resp_h=131.159.14.23, resp_p=22/tcp], F, 1.99, OpenSSH_3.9p1
    SSH banner, [orig_h=192.150.186.169, orig_p=49244/tcp, resp_h=131.159.14.23, resp_p=22/tcp], T, 2.0, OpenSSH_3.8.1p1

    real    0m22.037s
    user    0m21.720s
    sys     0m0.146s

    > time bro -r ssh-single-conn.trace ssh-banner.bro
    SSH banner, [orig_h=192.150.186.169, orig_p=49244/tcp, resp_h=131.159.14.23, resp_p=22/tcp], F, 1.99, OpenSSH_3.9p1
    SSH banner, [orig_h=192.150.186.169, orig_p=49244/tcp, resp_h=131.159.14.23, resp_p=22/tcp], T, 2.0, OpenSSH_3.8.1p1

    real    0m2.127s
    user    0m2.036s
    sys     0m0.073s

Not that both times shouldn't be decreased further, but generally it's
working.

TODO: The caching doesn't fully track dependencies yet. If source
files are modified directly, it should pick up on that and recompile,
but if they include other stuff recursively that gets changed, chances
are the caching will continue using an old version.