Changing the way one tells Bro what *.{pac2,evt} files to load.
The plugin used to load all *.{pac2,evt} from a preconfigured set of
directories, but that's quite inflexible. Instead one now gives the
list on the command line in addition to standard Bro scripts; Bro will
notice the extension and pass the files on to the HILTI plugin. If one
doesn't give an absolute path, the plugin will search them in its
pac2/ directory and further in all directories specified by the
optional BRO_PAC2_PATH environment variable (colon-separated, as
usual).

The *.evt language also gained a new keyword allowing to directly
specify a pac2 file to pull in with it, e.g., in ssh.evt:

    grammar ssh.pac2;

This will load ssh.pac2 automatically with ssh.evt, i.e., it's
sufficient now to list "ssh.evt" on the command line. The argument to
"grammar" is a path relative to the *.evt directory (and the normal
search paths if not found there); the pac2 extension can be skipped
and will then be added automatically.

Finally, it might even work to say "@load ssh.evt;" in a Bro script,
but that's untested ...