Initial infrastructure for an experimental Bro to HILTI compiler.
This will be my spare time project for a while. :) If anybody wants to
help, let me know; much of this is filling in gaps to support all of
the Bro language.

The current code doesn't do much yet but it can already
compile-and-jit a simple hello-world as well as call out to bif
functions (both Bro's original bifs as well as new bifs written using
the HILTI C API). Example:

    # cat foo.bro

    event bro_init() &priority = 42
        {
        print "Hello HILTI World!";
        print Hilti::is_compiled();
        print to_lower("FoObAr");
        }

    # bro -b ./foo.bro Hilti::compile_scripts=T
    Hello HILTI World!
    True
    foobar

Hilti::is_compiled() is a bif that returns a boolean indicating
whether the script code has been compiled. Note how Bro prints "True"
here, not "T": that's because HILTI prints booleans as True/False, not
as T/F as the standard interpreter would; take that as proof that it's
running compiled code. :-)

(Ignore the "&priority = 42"; that's a magic marker to select event
handlers to compile, because we can't compile them all yet.)