Tests for byte-to-uint/integer conversions
I am not sure which of both syntaxes is better (or if both do).

The tests are supposed to interpret the raw byte-value as a integer
or unsigned integer value either in little/bit endian. Truncating
to 64-bit for as_int/as_uint is ok from my point of view. With the
&parse syntax, this is a bit more explicitly visible.

In any case, it would be nice to easily be able to parse variable-width
bytestrams as integer values - which either of these tests would
accomplish.

Additionally for cases where the width is known before,
uint/int 24/48 types would be handy :)