test for skip-feature.
Skip is supposed to skip the next n bytes in the data stream without
storing them in any data structure at all (which is different from
just using &transient). This can be used when having a big data
stream that only contains little bits of interesting information that
are distributed in the stream. E.g. a tar-file where one is only interested
in parsing the tar file headers and not the actual file data.