Extending HILTI for-each loop with "magic" labels to change control flow.
Jump to @__LOOP_END will leave the loop, and jumping to @__LOOP_NEXT
will start the next iteration. This works as expected in nested loops
(i.e., the labels refer to the innermost loop).

Not the nicest way to add break and next constructs but works for now.
Adding corresponding instructions would be trivial but having labels
can actually be handy.