Ignore HILTI blocks that fail to return a value from a function if they can never be reached anyways.
Actually the new code isn't quite working yet and hence disabled: for
now, we don't report any error if a function doesn't return a value
even though it should. LLVM will complain anyways, however at that
level it's hard to figure out where the problem is.

To fix this we'll need to add reachibility analysis to the CFG (which
shouldn't be hard, we have all the information; it's just not there
yet).