Fixing a pretty fundamental flaw in HILTI's memory management.
Turns out the reference counting wasn't thread-safe, which led to race
conditions when passing objects between threads. The counter
operations are now atomic.

Unfortunately this could be pretty expensive I guess. Need to
investigate if there are ways to improve. For example, it's using the
the most restrictive consistency model for now; not sure if we could
relax that. Also, most objects will never be passed around between
threads and hence wouldn't need atomic operations - but I'm afraid we
can't figure that out automatically.