Adding statement management for connection records to compiler runtime.
We keep a mapping between Bro connection vals and HILTI connection
structs. When a Bro event gives us a val, we replace it with the HILTI
version, which then also tracks all the changes that a script is
doing. The one case this doesn't handle is changes the core is doing
to the connection record behind the scenes; but it should be good
enough for now. With this, the compiler can now handle
base/protocols/{dns,http}, and the compiled code generates the
corresponding {dns,http}.log files.

Plus a number of smaller fixes and test updates.