Very (very!) basic x509 parser in binpac++.
Parses the basic certificate data structures (and does nothing else).

Due to the fact that cross-module imports do not work correctly yet,
this is intermingled with the asn1-parsing and generally needs quite
a bit of cleanup.

But - it parses out all v1 certificate fields into a manageable data
structures. Adding extensions as we see fit would not be much of a problem.

Next step: forget about certificates and try to implement bacnet. The
data types we support should be sufficient for at least a few of the
messages :)

This probably abuses a few binpac++-features.