Redoing compilation of Bro's any type.
We now no longer pass it around as Val object all the time, even
inside HILTI code, but instead as a HILTI struct with a pointer the
corresponding HILTI value; to be converted to Val on demand only. This
fixes some problems with the former approach and also is more efficient.