Handle segmentation in BACnet.
This is not solved in an elegant way - but I do not really know
how to improve this with the current feature set of binpac.

Basically, for each of the 2 request types that support segmentation,
a map is created which is indexed by the invoke-ID. The bytes of all
packets until the last one are appended to the map. When the last
packet is encountered, the whole byte stream is parsed.

If we do not see the last packet of a specific segmented message, we
never delete the data and leak memory.