Instruction matching fixes.
We coercing argument to make them fit with an instruction, we know
require that at least one operand stays with the original type for
instructions that have more than one.

We also no longer consider whether an operand is a constant or not;
that doesn't really imply whether it can be passed as a non-constant
operand or not.