A HILTI "clone" operator for deep-copying values.
This currently works for all atomic types as well as for strings,
bytes, lists, vectors, maps, sets, regexps, files, tuples, callables,
and structs.

This comes with a bit of internal machinery that can also be used from
C directly. It deals correctly (I hope :) with reference cycles and
has an alternative (yet untested) mode for doing shallow copies
instead of deep.

The infrastructure is also prepared for cloning a value so that in can
then be used in a different thread (which by default is not
necessarily the case). That hasn't been tested at all yet, is not yet
used/exposed, and is also not yet fully implemeneted.