Support for adding custom HILTI code, including hook implementations that get triggered when Bro throws the correspondings events.
I call this "Lorenzo mode". :)

If one adds a *.hlt file to the command line, the Bro plugin will load
and compile that. If that code contains a hook that matches, in name
and type, one of the hooks one which the compiler generates for a Bro
event, that hook will execute along whenever Bro throws that event.

See bro/tests/compiler/custom-hilti-code-* for a couple examples.

This is somewhat preliminary right now. Deficiencies:

  (1) One needs to declare all the necessary types "by hand". Need to
      think about a way to make the compiler write them out so that
      the custom HILTI code can just import them.

  (2) There must be at least one actual Bro event handler defined for
      the custom hook to trigger, as otherwise Bro won't bother to
      raise the event in the first place.

  (3) It currently only works when script compilation is enabled
      (which is still limited and somewhat fragile). In principle the
      custom code could just as well trigger when the compiler is
      otherwise not used; however that needs some additional logic
      that isn't in place right now. (Adding this would likely also
      solve the previous point.)

I'm planing to take a look at (2) and (3) soon. (1) should be fine for
now.