A constructor expression for unions, along with Bro support to turn them into records.
These two now create instances of anonymous union types on the fly:

    (1) union($a=1, $b="string", $c=a.?b)
    (2) union(1, "string", a.?b)

The first creates a union with three fields of the given names; the
3rd field will remain unset if a.b is not set.

The second does the same but leaves the field names anonymous (to be
precise, it generates internal ones; you'll see them with "print").

When passing one of these to Bro via *.evt files they turn into
records, with fields left unset as on the BinPAC++ side.

TODO: Assigning these to variables of other union types should
eventually work as long as the field names/type match, but the
coercion code isn't implemented yet.