A new ".?" operator for BinPAC++.
Given a union A with a field B, A.?B returns the value of the field if
it's set and throws an exception AttributeNotSet if not.

This is mainly for use in specific evaluation contexts that catch that
exception to handle it differently in some way. In particular, there
will be a union ctor upcoming that will eave fields unset if their
init expressions throws this exception.

Note that just using "." would throw an exception as well, but a
different one that's not caught anywhere default and hence triggers an
error. Using this new operator makes it explicit that encountering an
unset field is an expected situation.

Not sure yet I like the ".?" syntax though. Other ideas welcome.