Deep-copy values when exchanging them between threads.
This has two parts: (1) thread.schedule now deep-copies all arguments
passed to the scheduled function call; and (2) channels now deep-copy
elements on write.

The motivation here is that, while this is more expensive than passing
referecences around, it fully isolates threads from each other, which
makes semantics much more crisp. Essentially there's now no shared
state anymore and one cannot accidentally access another thread's
objects. (There's one exception: channels themselves aren't
deep-copied when passed around; they now manage a shared data
structure themselves internally. After all they are *meant* to be
shared so that one can communicate).

This change also avoids the need for atomic reference counters, which
have been disabled.

Note that there's some potential for optimizations here: when one
passes an object that can be proven to not be accessed anymore in the
originating thread afterwards, we could switch back to passing a
reference instead. The existing liveness analysis should be able to do
that in many cases. Also, one could optimize the memory allocation so
that the deep-copies are performed within the memory space of the
target thread. Most of the infrastructure is there for that, but it's
not yet enabled.