Extending bytes to enable insertion of separator objects.
This introduces a new instruction "bytes.append_object n b o" that
appends an object o of arbitrary type to the end of a bytes instance
b. For normal bytes operations, this object will remain invisible and
appear just like an EOD. However, there's a separate set of new
object-related instructions that operate on such objects:

    bytes.at_object i [t]
        Test if an iterator i points to an object [of type t].

    bytes.retrieve_object i
        Extract object at iterator position i.

    bytes.skip_object i
        Move an iterator beyond an object.

Per above, for "normal" instructions an iterator pointing to an object
would just look line "end".

This is all still experimental, we'll have to see if the semantics are
right.