Union extensions and fixes, plus a struct extension.
We have union constructors now:

    type A = union {
        int<64> i,
        bool b,
        string test
    }

    type B = union<bool, string>

void run() {
    local A a
    local B b

    a = union<A>(test: "abc")
    b = union<B>("abc")
}

This actually works, too:

    a = union(test: "abc")
    b = union("abc")

The 2nd form create a union constant with just one field and then
coerces that on assignment.

Unions now also coerce to bool: true iff a value is set.

The commit also includes a few union fixes, plus a struct change:
fields can now be marked as "anonymous", meaning that when printed an
runtime their field name won't be shown. This can currently be set
internally on the AST node only, it's not exposed by the HILTI language.