Adding a BinPAC++ option to record for all fields their offsets inside the binary input data.
This is by default off, as it's potentially expensive (it triggers
more buffering and maintains two additional vectors inside the parse
objects). pac-driver/pac-dump/binpac++ get a new option -f to
activate. However, only pac-dump currently makes use of it: in JSON
mode, it includes the offsets into the output.