First version of DFD: Dynamic Format Detection.
There's a new sink method try_connect_mime_type() that works mostly
like connect_mime_type() but attaches all matching subunits in a new
preliminary trial mode, assuming that it's not clear yet if the sink's
input is indeed in their format. If in this mode a parse error occurs,
that's internally treated as a soft error that just leads to the
parser being disabled for the sink.

There are a few additional pieces to that scheme:

    - unit.confirm() confirms that a unit is seeing the right stuff,
      disabling trial mode. This can only be called on a top-level
      unit; that's not that great, but unclear if it could be
      supported down in the parse tree (worst case is one will alsways
      need to pass it down).

    - unit.disable(msg) explicitly disables parsing, just as a parse
      error in trial mode would. If this happens inside a unit that is
      directly instantiated by a host application (i.e, not connected
      via sink), it will propagate as a ParserDisabled exception
      upstream.

    - the hook %confirmed() runs when unit.confirm() is called. This
      hook can only be defined in a top-level unit.

    - the hook %disabled() runs when a parser gets disabled, either
      via parse error in trial mode or explicitly via disable(). This
      hook can only be defined in a top-level unit.

TODO: Much of this isn't tested yet.