Second version of DFD: now with %synchronize support.
If a unit connected with try_connect_mime_type() defines
synchronization support, that actviated first before the actual
parsing proceeds. That can (1) be more effificient (e.g.., if it can
quickly scan for a regexp, and (2) allows for parsing to start
midstream once the synchronization has identified a suitable spot
(this will allow, e.g., to reasonably parse partial connections). Once
synchronization completes, the actual parsing will stay in trial mode,
per the previous commit (i.e., we follow the 2-stage process of Bro's
DPD here).

In addition, this commits adds fixes and tests for the features from
the previous one.

TODO: Leaving an optimization for later: if multiple units define sync
support via regexp, we to do the comine-all-the-regexps-into-one
scheme for speeding up the matching.