pac-driver: Provide caplen to Hilti
Grammars often depend on the knowledge of the length of the content.
This is especially the case for Ethernet packets. Thus, the modified
pac-driver needs to provide this information to Hilti. The caplen
provided by libpcap as uint32_t is added at the very beginning to the
input of Hilti. In a grammar, this is the first data to be interpreted.
By default, a grammar expects big endian; thus, caplen is converted
accordingly.