pac-driver: Add possibility to read pcap files
Add a new command line option -f. The file is read using libpcap. This
shares a lot of code with listening on a network interface.

The grammar networkinterface.pac2 was renamed to libpcap.pac2 because it
is now used for both reading from pcap files and network devices.

There is now an ambiguity with parsers/libpcap.pac2 and
parsers/pcap.pac2, there should be a way to resolve this.