Hand over timestamp of packet capture to the grammar
When using libpcap to read from a network interface or a pcap file,
hand over the packet capture time and also the total length of the
packet; not only the captured length.

Currently, the timestamps are handled as being 64 bit. This has to be
generalized to cover platforms where time_t and suseconds_t are 32 bit,
as well.