Porting HILTI to LLVM 3.9.
This commit starts porting the code base over to LLVM 3.9 (and
dropping support for all earlier versions). In addition to adapting to
LLVM's API changes, the target is to run with an out-of-the-box
LLVM/clang installation, so that we no longer require everybody to
compile their own LLVM/clang in exactly the right way. Right now, this
work has been done for the HILTI and libhilti components; they are
working with a standard 3.9, and all their tests pass. BinPAC/Spicy
and the Bro plugin have not yet been ported over; these are completely
broken in this branch right now and will require more work to bring
back.

As part of the 3.9 overhaul, this commit also includes the following:

    - Adding custom RTTI infrastructure, so that we can compile
      without -fno-rtti (which LLVM does by default).

    - Cleaning up how sharing state between libhilti and JIT'ing host
      application works, making things quite a bit simpler. The
      additional libhilti-jit.a is gone.

    - Cleaning up the CompilerContxt API for compiling and linking.
      In particular, we now use unique_ptr<> consistently for passing
      modules around.

    - Cleaning up the CMake installation (still in progress).

    - Adding a new tool 'hiltip' that generates C prototypes for host
      applications. This functionalitoy has moved over from "hiltic
      -P" into "hiltip". With that, the kludgy "hiltic-nojit" is now
      gone.

    - Running the code through clang-format; there's a new Makefile
      target for that ("make clang-format"). However, clang-format
      needs a patch currently to support a new option that our
      clang-format configuration requires. (The reformatting messes up
      git history unfortuantely, sorry!).

    - Lots of little updates and cleanups.